treewide: fix some warnings found with GCC 7

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -57,7 +57,7 @@
 
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
-	int t = readl_relaxed(addr); \
+	u64 t = readl_relaxed(addr); \
 	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
 			(mask)) << (shift)), addr); \
 } while (0)
